:hammer: split AppSourceIcon and SideAppSourceIcon transformer pipelines
Introduce DefaultAppSourceIconTransformer (trim-only, used by AppSourceIcon)
and SideAppSourceIconTransformer (trim + fixed-ratio crop, used by
SideAppSourceIcon). Extract shared alpha-detection and subset-copy logic into
AppSourceIconAlphaTrim to deduplicate the two transformations.

Lower the alpha threshold (12 -> 4) and raise the safety margin (1.5% ->
2.5%) so anti-aliased edge pixels survive the trim, and oversample the trim
request by 2x so downsampling to the display size stays smooth.
